# crates/server/openapi.yaml
...
openapi: 3.0.3
info:
  title: Remote Harness Control Plane API
  version: 0.1.0
  description: |
    Control plane REST API. Human-readable API prose and behavior notes live in
    `docs/API_OVERVIEW.md`; this spec is the machine-readable contract for codegen and review.
servers:
  - url: /
    description: Relative to configured control plane base URL
...
paths:
...
  /agents/{id}/inbox:
    get:
      tags: [agents]
      summary: List pending inbox tasks for an agent
      operationId: getAgentInbox
      security:
        - bearerApiKey: []
        - headerApiKey: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Agent id (same as session params.agent_id for workflow inbox)
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: cursor
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Paginated pending inbox tasks
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedInboxTaskItems"
        "400":
          description: Invalid cursor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StandardError"
        "401":
          description: Missing or invalid API key
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StandardError"
        "404":
          description: Agent not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StandardError"
        "503":
          description: Database not configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StandardError"
    post:
      tags: [agents]
      summary: Enqueue a task to an agent inbox
      operationId: postAgentInbox
      security:
        - bearerApiKey: []
        - headerApiKey: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PostAgentInboxRequest"
      responses:
        "202":
          description: Task accepted into inbox queue
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PostAgentInboxResponse"
        "400":
          description: Invalid body
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StandardError"
        "401":
          description: Missing or invalid API key
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StandardError"
        "404":
          description: Agent not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StandardError"
        "409":
          description: No running inbox session or conflict
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StandardError"
        "503":
          description: Database not configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StandardError"
...
components:
  securitySchemes:
    bearerApiKey:
      type: http
      scheme: bearer
      description: "API key as Bearer token (`Authorization: Bearer <api_key>`)"
    headerApiKey:
      type: apiKey
      in: header
      name: X-API-Key
      description: "API key in `X-API-Key` header"
  schemas:
...
    StandardError:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              example: not_found
            message:
              type: string
              example: Resource not found
            details:
              type: object
              additionalProperties: true
...
    PostAgentInboxRequest:
      type: object
      required: [payload]
      properties:
        payload:
          type: object
          additionalProperties: true
          description: Opaque JSON; must include non-empty string field `message` or `prompt` for worker prompt extraction
        persona_id:
          type: string
          nullable: true
    PostAgentInboxResponse:
      type: object
      required: [task_id]
      properties:
        task_id:
          type: string
          format: uuid
          description: Inbox queue row id until promoted to a job on worker pull
    InboxTaskItem:
      type: object
      required: [task_id, payload, enqueued_at]
      properties:
        task_id:
          type: string
          format: uuid
        payload:
          type: object
          additionalProperties: true
        enqueued_at:
          type: string
          format: date-time
    PaginatedInboxTaskItems:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/InboxTaskItem"
        next_cursor:
          type: string
          nullable: true
...